Refactor test_dht_node.py and related files
